index depends on tilejson route, move it into factory
