+++ timvt/endpoints/factory.py
@@ -1,6 +1,6 @@
 """TiVTiler.endpoints.factory: router factories."""
 from dataclasses import dataclass, field
-from typing import Any, Callable, Dict, Optional, Type
+from typing import Any, Callable, Dict, List, Optional, Type
 
 from asyncpg.pool import Pool
 from morecantile import TileMatrixSet
@@ -55,6 +55,7 @@ class VectorTilerFactory:
         """Register Tiler Routes."""
         self.tile()
         self.tilejson()
+        self.index()
 
     def url_for(self, request: Request, name: str, **path_params: Any) -> str:
         """Return full url (with prefix) for a specific endpoint."""
@@ -63,6 +64,27 @@ class VectorTilerFactory:
         if self.router_prefix:
             base_url += self.router_prefix.lstrip("/")
         return url_path.make_absolute_url(base_url=base_url)
+
+    ############################################################################
+    # /index
+    ############################################################################
+    def index(self):
+        """Register /index endpoint"""
+
+        @self.router.get(
+            "/index",
+            response_model=List[TableMetadata],
+            description="Return available tables.",
+            deprecated=True,
+        )
+        async def display_index(request: Request):
+            """Return JSON with available table metadata. """
+            return [
+                TableMetadata(
+                    **table, link=self.url_for(request, "tilejson", table=table["id"])
+                )
+                for table in request.app.state.Catalog
+            ]
 
     ############################################################################
     # /tiles
